+++ xarray/tests/test_computation.py
@@ -9,7 +9,7 @@ import pytest
 
 import xarray as xr
 from xarray.core.computation import (
-    _UFuncSignature, broadcast_compat_data, collect_dict_values,
+    _UFuncSignature, result_name, broadcast_compat_data, collect_dict_values,
     join_dict_keys, ordered_set_intersection, ordered_set_union,
     unified_dim_sizes, apply_ufunc)
 
@@ -36,6 +36,19 @@ def test_signature_properties():
     assert _UFuncSignature([['x']]) != _UFuncSignature([['y']])
 
 
+def test_result_name():
+
+    class Named(object):
+        def __init__(self, name=None):
+            self.name = name
+
+    assert result_name([1, 2]) is None
+    assert result_name([Named()]) is None
+    assert result_name([Named('foo'), 2]) == 'foo'
+    assert result_name([Named('foo'), Named('bar')]) is None
+    assert result_name([Named('foo'), Named()]) is None
+
+
 def test_ordered_set_union():
     assert list(ordered_set_union([[1, 2]])) == [1, 2]
     assert list(ordered_set_union([[1, 2], [2, 1]])) == [1, 2]
@@ -55,6 +68,8 @@ def test_join_dict_keys():
     assert list(join_dict_keys(dicts, 'right')) == ['y', 'z']
     assert list(join_dict_keys(dicts, 'inner')) == ['y']
     assert list(join_dict_keys(dicts, 'outer')) == ['x', 'y', 'z']
+    with pytest.raises(ValueError):
+        join_dict_keys(dicts, 'exact')
     with pytest.raises(KeyError):
         join_dict_keys(dicts, 'foobar')